+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.9.0
+
+- Added windows support
+- Initial template path fixed
+- Path saperator depends on Platform added to 
+- Better Cli Experience
+- Skipped copying 'name.txt', 'instructions.md' and 'config.yaml'
+- Instructions printing added
+
 ## 0.8.2
 
 - Version print issue fixed